+++ tests/test_application.py
@@ -4,8 +4,12 @@
 Copyright (C) 2022, Auto Trader UK
 
 """
+from typing import Union
+
 import numpy as np
+import numpy.typing as npt
 import pandas as pd
+import pytest as pytest
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from mlflow.pyfunc import PyFuncModel  # type: ignore
@@ -35,11 +39,22 @@ def test_build_app_has_predictions_endpoint(pyfunc_model: PyFuncModel):
     assert response.status_code != 404
 
 
+@pytest.mark.parametrize(
+    "pyfunc_output_type",
+    ["ndarray", "series", "dataframe"],
+)
 def test_build_app_returns_good_predictions(
-    pyfunc_model: PyFuncModel,
     model_input: pd.DataFrame,
-    model_output: np.typing.ArrayLike,
+    pyfunc_output_type: str,
+    request: pytest.FixtureRequest,
 ):
+    pyfunc_model: PyFuncModel = request.getfixturevalue(
+        f"pyfunc_model_{pyfunc_output_type}"
+    )
+    model_output: Union[npt.ArrayLike, pd.DataFrame] = request.getfixturevalue(
+        f"model_output_{pyfunc_output_type}"
+    )
+
     app = build_app(pyfunc_model)
 
     client = TestClient(app)
@@ -47,6 +62,10 @@ def test_build_app_returns_good_predictions(
     request_data = f'{{"data": {df_str}}}'
     response = client.post("/predictions", data=request_data)
     assert response.status_code == 200
-    assert response.json()["data"] == [
-        {"prediction": v} for v in np.nditer(model_output)
-    ]
+    results = response.json()["data"]
+    try:
+        assert model_output.to_dict(orient="records") == results  # type: ignore
+    except (AttributeError, TypeError):
+        assert [
+            {"prediction": v} for v in np.nditer(model_output)
+        ] == results  # type: ignore